Add routes file and remove global types file
