+++ src/components/authentication/CreateAccountPage.tsx
@@ -1,6 +1,16 @@
 import React from "react";
 import { Text } from "react-native";
 
-export default function CreateAccountPage() {
+import {
+  CreateAccountPageRouteProp,
+  CreateAccountPageNavigationProp,
+} from "../../routes";
+
+type CreateAccountPageProps = {
+  route: CreateAccountPageRouteProp;
+  navigation: CreateAccountPageNavigationProp;
+};
+
+export default function CreateAccountPage({}: CreateAccountPageProps) {
   return <Text>Hello world</Text>;
 }
